+++ packages/rrweb/record/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rrweb/record
 
+## 0.0.73
+
+### Patch Changes
+
+- Updated dependencies [[`f95a0ec`](https://github.com/PostHog/posthog-js/commit/f95a0ec68270bf9116d29875733c1a43e9b91331)]:
+    - @posthog/rrweb-types@0.0.62
+    - @posthog/rrweb@0.0.73
+
 ## 0.0.72
 
 ### Patch Changes